refactor OrderCard and ExchangeOrderCard
