+++ ui/src/components/Broker/BrokerOrders.tsx
@@ -3,20 +3,24 @@ import { Button, Card, Form, Header } from 'semantic-ui-react'
 
 import { useLedger, useStreamQuery } from '@daml/react'
 import { Order, BrokerOrderRequest, BrokerOrder } from '@daml.js/da-marketplace/lib/Marketplace/Trading'
+import { ContractId } from '@daml/types'
 
 import { ExchangeIcon, OrdersIcon } from '../../icons/Icons'
 import { unwrapDamlTuple, wrapDamlTuple } from '../common/damlTypes'
 import { parseError, ErrorMessage } from '../common/errorTypes'
 import FormErrorHandled from '../common/FormErrorHandled'
+import ExchangeOrderCard from '../common/ExchangeOrderCard'
 import Page from '../common/Page'
-import { ContractId } from '@daml/types'
+
 
 import './BrokerOrders.css'
+
 
 type Props = {
     sideNav: React.ReactElement;
     onLogout: () => void;
-}
+};
+
 
 const BrokerOrders: React.FC<Props> = ({ sideNav, onLogout }) => {
     const allExchangeOrders = useStreamQuery(Order).contracts;
@@ -37,81 +41,15 @@ const BrokerOrders: React.FC<Props> = ({ sideNav, onLogout }) => {
                 { allBrokerOrders.map(o => <BrokerOrderCard key={o.contractId} cdata={o.payload}/>)}
 
                 <Header as='h4'>Exchange Orders</Header>
-                { allExchangeOrders.map(o => <OpenOrderCard key={o.contractId} order={o.payload}/>)}
+                { allExchangeOrders.map(o => <ExchangeOrderCard key={o.contractId} order={o.payload}/>)}
 
             </div>
         </Page>
     )
-}
-
-type BrokerOrderCardProps = {
-    cdata: BrokerOrder;
 };
 
-const BrokerOrderCard: React.FC<BrokerOrderCardProps> = (props) => {
-    const [ loading, setLoading ] = useState(false);
-    const base = unwrapDamlTuple(props.cdata.pair)[0].label;
-    const quote = unwrapDamlTuple(props.cdata.pair)[1].label;
-    const label = props.cdata.isBid ? `Buy ${base}/${quote}` : `Sell ${base}/${quote}`;
-    const amount = props.cdata.isBid ? `+ ${props.cdata.qty} ${base}` : `- ${props.cdata.qty} ${base}`;
-    const price = `${props.cdata.price} ${quote}`;
 
-    const customer = props.cdata.brokerCustomer;
-
-    const [ depositCid, setDepositCid ] = useState<string>('');
-    const [ error, setError ] = useState<ErrorMessage>();
-    const ledger = useLedger();
-
-    const handleAcceptBrokerOrderFill = async (event: React.FormEvent) => {
-        event.preventDefault();
-        setLoading(true);
-        try {
-            const key = wrapDamlTuple([props.cdata.broker, props.cdata.brokerOrderId])
-            const args = {
-                depositCid
-            }
-            await ledger.exerciseByKey(BrokerOrder.BrokerOrder_Fill, key, args);
-        } catch (err) {
-            setError(parseError(err));
-        }
-        setLoading(false);
-    }
-
-    return (
-        <FormErrorHandled
-            className='order-card-container'
-            error={error}
-            clearError={() => setError(undefined)}
-        >
-        <div className='order-card-container'>
-            <div className='order-card'>
-                <Card fluid className='order-info'>
-                    <div><ExchangeIcon/> {label}</div>
-                    <div>{ amount }</div>
-                    <div>{`@ ${price}`}</div>
-                    <div>{`customer: ${customer}`}</div>
-                    <div>{`id: ${props.cdata.brokerOrderId}`}</div>
-                </Card>
-
-                <Form.Input
-                    className='orderid-input'
-                    placeholder='depositCid'
-                    value={depositCid}
-                    onChange={e => setDepositCid(e.currentTarget.value)}
-                />
-                <Button
-                    basic
-                    color='black'
-                    content='Fill Order'
-                    loading={loading}
-                    onClick={handleAcceptBrokerOrderFill}
-                />
-            </div>
-        </div>
-        </FormErrorHandled>
-    )
-};
-
+// Broker Order Request
 type BrokerOrderRequestCardProps = {
     cid: ContractId<BrokerOrderRequest>;
     cdata: BrokerOrderRequest;
@@ -120,8 +58,9 @@ type BrokerOrderRequestCardProps = {
 
 const BrokerOrderRequestCard: React.FC<BrokerOrderRequestCardProps> = ({children, ...props}) => {
     const [ loading, setLoading ] = useState(false);
-    const base = unwrapDamlTuple(props.cdata.pair)[0].label;
-    const quote = unwrapDamlTuple(props.cdata.pair)[1].label;
+    const pair = unwrapDamlTuple(props.cdata.pair)
+    const base = pair[0].label;
+    const quote = pair[1].label;
     const label = props.cdata.isBid ? `Buy ${base}/${quote}` : `Sell ${base}/${quote}`;
     const amount = props.cdata.isBid ? `+ ${props.cdata.qty} ${base}` : `- ${props.cdata.qty} ${base}`;
     const customer = props.cdata.brokerCustomer;
@@ -161,61 +100,58 @@ const BrokerOrderRequestCard: React.FC<BrokerOrderRequestCardProps> = ({children
                     <div>{`deposit: ${depositCid.substr(depositCid.length - 8)}`}</div>
                 </Card>
 
-                <Form.Input
-                    className='orderid-input'
-                    placeholder='id'
-                    value={brokerOrderId}
-                    onChange={e => setBrokerOrderId(e.currentTarget.value)}
-                />
-                <Button
-                    basic
-                    color='black'
-                    content='Accept Order'
-                    loading={loading}
-                    onClick={handleAcceptBrokerOrderRequest}
-                />
+                <Form.Group>
+                    <Form.Input
+                        className='orderid-input'
+                        placeholder='id'
+                        value={brokerOrderId}
+                        onChange={e => setBrokerOrderId(e.currentTarget.value)}
+                    />
+                    <Button
+                        basic
+                        color='black'
+                        content='Accept Order'
+                        loading={loading}
+                        onClick={handleAcceptBrokerOrderRequest}
+                    />
+                </Form.Group>
             </div>
         </div>
         </FormErrorHandled>
     )
-}
+};
 
 
+// Broker Order
+type BrokerOrderCardProps = {
+    cdata: BrokerOrder;
+};
 
-type OrderProps = {
-    order: Order;
-}
 
-const OrderCard: React.FC<OrderProps> = ({ children, order }) => {
-    const base = unwrapDamlTuple(order.pair)[0].label;
-    const label = order.isBid ? `Buy ${base}` : `Sell ${base}`;
-    const amount = order.isBid ? `+ ${order.qty} ${base}` : `- ${order.qty} ${base}`;
-
-    return (
-        <div className='order-card-container'>
-            <div className='order-card'>
-                <Card fluid className='order-info'>
-                    <div><ExchangeIcon/> {label}</div>
-                    <div>{ amount }</div>
-                </Card>
-
-                { children }
-            </div>
-        </div>
-    )
-}
-
-const OpenOrderCard: React.FC<OrderProps> = ({ order }) => {
+const BrokerOrderCard: React.FC<BrokerOrderCardProps> = (props) => {
     const [ loading, setLoading ] = useState(false);
+    const pair = unwrapDamlTuple(props.cdata.pair)
+    const base = pair[0].label;
+    const quote = pair[1].label;
+    const label = props.cdata.isBid ? `Buy ${base}/${quote}` : `Sell ${base}/${quote}`;
+    const amount = props.cdata.isBid ? `+ ${props.cdata.qty} ${base}` : `- ${props.cdata.qty} ${base}`;
+    const price = `${props.cdata.price} ${quote}`;
+
+    const customer = props.cdata.brokerCustomer;
+
+    const [ depositCid, setDepositCid ] = useState<string>('');
     const [ error, setError ] = useState<ErrorMessage>();
     const ledger = useLedger();
 
-    const handleCancelOrder = async (event: React.FormEvent) => {
+    const handleAcceptBrokerOrderFill = async (event: React.FormEvent) => {
         event.preventDefault();
         setLoading(true);
         try {
-            const key = wrapDamlTuple([order.exchange, order.orderId])
-            await ledger.exerciseByKey(Order.Order_RequestCancel, key, {});
+            const key = wrapDamlTuple([props.cdata.broker, props.cdata.brokerOrderId])
+            const args = {
+                depositCid
+            }
+            await ledger.exerciseByKey(BrokerOrder.BrokerOrder_Fill, key, args);
         } catch (err) {
             setError(parseError(err));
         }
@@ -228,17 +164,35 @@ const OpenOrderCard: React.FC<OrderProps> = ({ order }) => {
             error={error}
             clearError={() => setError(undefined)}
         >
-            <OrderCard order={order}>
-                <Button
-                    basic
-                    color='black'
-                    content='Cancel'
-                    className='basic-button-fill'
-                    loading={loading}
-                    onClick={handleCancelOrder}/>
-            </OrderCard>
+            <div className='order-card-container'>
+                <div className='order-card'>
+                    <Card fluid className='order-info'>
+                        <div><ExchangeIcon/> {label}</div>
+                        <div>{ amount }</div>
+                        <div>{`@ ${price}`}</div>
+                        <div>{`customer: ${customer}`}</div>
+                        <div>{`id: ${props.cdata.brokerOrderId}`}</div>
+                    </Card>
+
+                    <Form.Group>
+                        <Form.Input
+                            className='orderid-input'
+                            placeholder='depositCid'
+                            value={depositCid}
+                            onChange={e => setDepositCid(e.currentTarget.value)}
+                        />
+                        <Button
+                            basic
+                            color='black'
+                            content='Fill Order'
+                            loading={loading}
+                            onClick={handleAcceptBrokerOrderFill}
+                        />
+                    </Form.Group>
+                </div>
+            </div>
         </FormErrorHandled>
     )
-}
+};
 
 export default BrokerOrders;
